git-svn-id: https://svn.apache.org/repos/asf/sis/trunk@1724531 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-referencing/src/main/java/org/apache/sis/referencing/operation/transform/AbstractMathTransform.java
@@ -358,8 +358,8 @@ public abstract class AbstractMathTransform extends FormattableObject
      * transforming many points. The supplied array of ordinal values will contain packed ordinal values.
      *
      * <div class="note"><b>Example:</b> if the source dimension is 3, then the ordinates will be packed in this order:
-     * (<var>x<sub>0</sub></var>,<var>y<sub>0</sub></var>,<var>z<sub>0</sub></var>,
-     *  <var>x<sub>1</sub></var>,<var>y<sub>1</sub></var>,<var>z<sub>1</sub></var> …).
+     * (<var>x₀</var>,<var>y₀</var>,<var>z₀</var>,
+     *  <var>x₁</var>,<var>y₁</var>,<var>z₁</var> …).
      * </div>
      *
      * The default implementation invokes {@link #transform(double[], int, double[], int, boolean)} in a loop,
@@ -568,7 +568,7 @@ public abstract class AbstractMathTransform extends FormattableObject
                 }
             }
             for (int i=0; i<srcStop; i++) {
-                buffer[bufferedSrcOff + i] = (double) srcPts[srcOff + i];
+                buffer[bufferedSrcOff + i] = srcPts[srcOff + i];
             }
             assert !IterationStrategy.suggest(bufferedSrcOff, dimSource, 0, dimTarget, numBufferedPts).needBuffer;
             try {
@@ -709,7 +709,7 @@ public abstract class AbstractMathTransform extends FormattableObject
                 dstLength = numPts * dimTarget;
             }
             for (int i=0; i<srcLength; i++) {
-                buffer[i] = (double) srcPts[srcOff++];
+                buffer[i] = srcPts[srcOff++];
             }
             try {
                 transform(buffer, 0, dstPts, dstOff, numBufferedPts);
